+++ src/api/api_internal.h
@@ -1,18 +1,14 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+// Copyright (C) 2026 Antoine Boquet
+
 #ifndef MORPHEUS_API_INTERNAL_H
 #define MORPHEUS_API_INTERNAL_H
 #include <morpheus/morpheus.h>
 #include "../morphlib/runtime_context.h"
-struct morpheus_compat_output {
-  char *data;
-  size_t length;
-  size_t analysis_count;
-  size_t lemma_count;
-};
 struct morpheus_result {
   size_t count;
   morpheus_analysis *analyses;
   morpheus_truncated_fields *truncated_fields;
-  uint8_t *all_morph_flags;
 };
 morpheus_result *morpheus_result_create(size_t count);
 static inline morpheus_status morpheus_runtime_status(