+++ src/api/context.c
@@ -1,3 +1,6 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+// Copyright (C) 2026 Antoine Boquet
+
 #include <morpheus/morpheus.h>
 #include <stdio.h>
 #include <stdlib.h>