+++ src/compat/compat.c
@@ -1,8 +1,10 @@
+/* SPDX-License-Identifier: MPL-2.0 */
+
 #if !defined(_POSIX_C_SOURCE)
 #define _POSIX_C_SOURCE 200809L
 #endif
 
-#include "api_internal.h"
+#include <morpheus/compat.h>
 
 #include <stdio.h>
 #include <stdlib.h>
@@ -10,6 +12,23 @@
 
 #include "../anal/anal_internal.h"
 #include "../morphlib/runtime_context.h"
+
+struct morpheus_compat_output {
+  char *data;
+  size_t length;
+  size_t analysis_count;
+  size_t lemma_count;
+};
+
+static morpheus_status compat_runtime_status(
+    const morpheus_runtime_context *context)
+{
+  switch(morpheus_runtime_context_error(context)) {
+  case MORPHEUS_RUNTIME_ERROR_NO_MEMORY: return(MORPHEUS_NO_MEMORY);
+  case MORPHEUS_RUNTIME_ERROR_INTERNAL: return(MORPHEUS_INTERNAL_ERROR);
+  default: return(MORPHEUS_OK);
+  }
+}
 
 morpheus_status
 morpheus_compat_analyze(
@@ -41,7 +60,7 @@ morpheus_compat_analyze(
   previous=morpheus_runtime_context_activate(context);
   morpheus_runtime_context_clear_error(context);
   word=morpheus_check_word(input,(PrntFlags)flags);
-  runtime_status=morpheus_runtime_status(context);
+  runtime_status=compat_runtime_status(context);
   if(runtime_status != MORPHEUS_OK) {
     if(word) FreeGkword(word);
     morpheus_runtime_context_activate(previous);
@@ -61,7 +80,7 @@ morpheus_compat_analyze(
     PrntAnalyses(word,(PrntFlags)flags,stream);
     if(anal_buf() && anal_buf()[0]) fputs(anal_buf(),stream);
   }
-  runtime_status=morpheus_runtime_status(context);
+  runtime_status=compat_runtime_status(context);
   FreeGkword(word);
   morpheus_runtime_context_activate(previous);
   if(runtime_status != MORPHEUS_OK) {